requirements.txt
streamlit>=1.37
langchain>=0.2
langchain-community>=0.2
sentence-transformers>=3.0
faiss-cpu>=1.8
PyPDF2>=3.0
python-docx>=1.1
numpy>=1.26
faster-whisper>=1.0
transformers>=4.43
accelerate>=0.33
torch>=2.3
torchaudio>=2.3
pytest>=8.0
